# smartender-global/smartender-global-backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Description of the global smartender API"
  version: "1.0.0"
  title: "smartender-global-server"
host: "vahlkamp.selfhost.eu:10000"
basePath: "/smartender/api"
...
schemes:
- "https"
- "http"
security:
- password: []
paths:
...
  /drinks:
    get:
      tags:
      - "data"
      description: "gets drinks limited by paging"
      operationId: "drinksGET"
      parameters:
      - name: "offset"
        in: "query"
        description: "page-number, starting at 0"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Amount of items retrieve\ndefault 20, \nmin 1,\nmax 50\n"
        required: false
        type: "integer"
      responses:
        200:
          description: "successfully retrieved page of drinks"
          schema:
            $ref: "#/definitions/DrinkPage"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security: []
      x-swagger-router-controller: "Data"
  /drinks/{id}:
    get:
      tags:
      - "data"
      description: "returns the drink for the specified id"
      operationId: "drinksIdGET"
      parameters:
      - name: "id"
        in: "path"
        description: "an id of a drink"
        required: true
        type: "integer"
      responses:
        200:
          description: "successfully retrieved the drink"
          schema:
            $ref: "#/definitions/Drink"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security: []
      x-swagger-router-controller: "Data"
  /drinks/new:
    put:
      tags:
      - "editing"
      description: "creates a new drink"
      operationId: "drinksNewPUT"
      parameters:
      - in: "body"
        name: "drink"
        description: "drink to be created"
        required: true
        schema:
          $ref: "#/definitions/Drink"
      responses:
        201:
          description: "drink successfully created"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Editing"
  /drinks/update:
    put:
      tags:
      - "editing"
      description: "updates an existing drink"
      operationId: "drinksUpdatePUT"
      parameters:
      - in: "body"
        name: "drink"
        description: "updated drink with old id"
        required: true
        schema:
          $ref: "#/definitions/Drink"
      responses:
        201:
          description: "drink successfully updated"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Editing"
  /drinks/delete/{id}:
    delete:
      tags:
      - "editing"
      description: "deletes an existing drink"
      operationId: "drinksDeleteIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        description: "id of drink to be deleted"
        required: true
        type: "integer"
      responses:
        202:
          description: "drink successfully deleted"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Editing"
  /ingredient:
    get:
      tags:
      - "data"
      description: "gets ingredient limited by paging"
      operationId: "ingredientGET"
      parameters:
      - name: "offset"
        in: "query"
        description: "page-number, starting at 0"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Amount of items retrieve\ndefault 20, \nmin 1,\nmax 50\n"
        required: false
        type: "integer"
      responses:
        200:
          description: "successfully retrieved page of ingredients"
          schema:
            $ref: "#/definitions/IngredientPage"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security: []
      x-swagger-router-controller: "Data"
  /ingredient/{id}:
    get:
      tags:
      - "data"
      description: "returns the ingredient for the specified id"
      operationId: "ingredientIdGET"
      parameters:
      - name: "id"
        in: "path"
        description: "an id of an ingredient"
        required: true
        type: "integer"
      responses:
        200:
          description: "successfully retrieved the ingredient"
          schema:
            $ref: "#/definitions/Ingredient"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security: []
      x-swagger-router-controller: "Data"
  /ingredient/new:
    put:
      tags:
      - "editing"
      description: "creates a new ingredient"
      operationId: "ingredientNewPUT"
      parameters:
      - in: "body"
        name: "ingredient"
        description: "ingredient to be creates"
        required: true
        schema:
          $ref: "#/definitions/Ingredient"
      responses:
        201:
          description: "ingredient successfully created"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Editing"
  /ingredient/update:
    put:
      tags:
      - "editing"
      description: "updates an exting ingredient"
      operationId: "ingredientUpdatePUT"
      parameters:
      - in: "body"
        name: "ingredient"
        description: "updated ingredient with old id"
        required: true
        schema:
          $ref: "#/definitions/Ingredient"
      responses:
        201:
          description: "ingredient successfully updated"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Editing"
...
  /user/machines:
    get:
      tags:
      - "data"
      operationId: "userMachinesGET"
      parameters: []
      responses:
        200:
          description: "returned an array of machines that belong to the logged-in\
            \ user"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Machine"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Data"
...
securityDefinitions:
  password:
    type: "basic"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      alias:
        type: "string"
        description: "the users display-name"
    example:
      alias: "alias"
      id: 0
  Drink:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "id of the drink"
      name:
        type: "string"
        description: "name of the drink"
      Ingredients:
        type: "array"
        items:
          $ref: "#/definitions/Ingredient"
    example:
      Ingredients:
      - name: "name"
        parts: 5
        id: 6
        alcvol: 1
      - name: "name"
        parts: 5
        id: 6
        alcvol: 1
      name: "name"
      id: 0
  Ingredient:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "id of the ingredient"
      name:
        type: "string"
        description: "name of the ingredient"
      alcvol:
        type: "integer"
        description: "alcohol vol. as a percentage"
      parts:
        type: "integer"
        description: "describes, if in a drink, how many parts of that drink are this\
          \ ingredient"
    example:
      name: "name"
      parts: 5
      id: 6
      alcvol: 1
  DrinkPage:
    type: "object"
    properties:
      drinks:
        type: "array"
        description: "array with result items"
        items:
          $ref: "#/definitions/Drink"
      next:
        type: "string"
        description: "link to retrieve next page if possible"
      previous:
        type: "string"
        description: "link to retrieve previous page if possible"
    description: "a paging object that contains drinks"
    example:
      next: "next"
      previous: "previous"
      drinks:
      - Ingredients:
        - name: "name"
          parts: 5
          id: 6
          alcvol: 1
        - name: "name"
          parts: 5
          id: 6
          alcvol: 1
        name: "name"
        id: 0
      - Ingredients:
        - name: "name"
          parts: 5
          id: 6
          alcvol: 1
        - name: "name"
          parts: 5
          id: 6
          alcvol: 1
        name: "name"
        id: 0
  IngredientPage:
    type: "object"
    properties:
      ingredients:
        type: "array"
        description: "array with result items"
        items:
          $ref: "#/definitions/Ingredient"
      next:
        type: "string"
        description: "link to retrieve next page if possible"
      previous:
        type: "string"
        description: "link to retrieve previous page if possible"
    description: "a paging object that contains ingredients"
    example:
      next: "next"
      previous: "previous"
      ingredients:
      - name: "name"
        parts: 5
        id: 6
        alcvol: 1
      - name: "name"
        parts: 5
        id: 6
        alcvol: 1
  Machine:
    type: "object"
    properties:
      name:
        type: "string"
        description: "name of the machine"
      owner:
        $ref: "#/definitions/User"
      id:
        type: "integer"
        description: "unique id of this machine"
    description: "a machine"
    example:
      owner:
        alias: "alias"
        id: 0
      name: "name"
      id: 6
...
  inline_response_401:
    properties:
      required_scopes:
        type: "array"
        description: "scopes that are required for this operation"
        items:
          type: "string"
  inline_response_default:
    properties:
      message:
        type: "string"
...